Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/active.https-expected.txt


PASS active is set 
PASS The ServiceWorker objects returned from active attribute getter that represent the same service worker are the same objects 
